% Generated by roxygen2 (4.0.2): do not edit by hand
\name{build_outlier_combinations}
\alias{build_outlier_combinations}
\title{Build outlier combinations}
\usage{
build_outlier_combinations(z)
}
\arguments{
\item{z}{a vector of sorted z-scores}
}
\value{
a list of binary vectors. Each vector represents an outlier model
 where outliers are denoted by 1, otherwise 0.
}
\description{
Build all possible combinations of over- and under-expressed outliers.
}
\details{
Generate all possible combinations of outliers, starting from the top and
 bottom of the sorted input.
}
\seealso{
\code{\link{find_tissue_outliers}}
}

