% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find_tissue_outliers}
\alias{find_tissue_outliers}
\title{Find over- or under-expressed tissue-specific genes}
\usage{
find_tissue_outliers(x, epsilon = 1e-04)
}
\arguments{
\item{x}{a vector of expression values}

\item{epsilon}{a very small number for offsetting zero values}
}
\value{
a binary vector representing the ouliers (1) and non-outliers (0)
}
\description{
The third step in ROKU algorithm. Identify tissue-specific outliers.
}
\details{
See the \href{http://www.biomedcentral.com/1471-2105/7/294#sec4}{Methods} from
the manuscript for more details.

Identifiers over- or under-expressed tissue-specific genes using
the U statistic from
\href{http://www.ncbi.nlm.nih.gov/pubmed/12499447}{Kadota et al. (2003)}. See
\code{\link{u_statistic}} for more details.

The combination of outliers that where \emph{U} is minimized is returned.

If no outliers are found, then k-means clustering is performed to cluster
the expression values.
}
\references{
Kadota et al (2006). ROKU: a novel method for identification of tissue-specific genes. BMC Genomics, 7:294.
\url{http://www.biomedcentral.com/1471-2105/7/294}

Kadota et al (2003). Detection of genes with tissue-specific expression patterns using Akaike's information criterion procedure. Physiol Genomics, 12(3):251-9. \url{http://www.ncbi.nlm.nih.gov/pubmed/12499447}
}
\seealso{
\code{\link{u_statistic}} for calculation of \emph{U}
\code{\link{build_outlier_combinations}} for building all possible outlier models
}

