% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{psidata}
\alias{psidata}
\title{Sample genes with simulated percent spliced-in (PSI) data}
\format{A 8 x 10 data frame}
\usage{
psidata
}
\description{
Contains simulated PSI data to test for tissue-specificity.
}
\keyword{datasets}

