% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ROKU_as_df}
\alias{ROKU_as_df}
\title{Convert ROKU results as data frame}
\usage{
ROKU_as_df(r, meta = NULL)
}
\arguments{
\item{r}{}

\item{meta}{}
}
\value{
a data frame
}
\description{
Take ROKU list and convert to data frame
}
\seealso{
\code{\link{ROKU}}
}

