% Generated by roxygen2 (4.0.2): do not edit by hand
\name{preprocess_with_tukey}
\alias{preprocess_with_tukey}
\title{Normalize by Tukey biweight}
\usage{
preprocess_with_tukey(x)
}
\arguments{
\item{x}{A vector of length \emph{m} where \emph{m} is the number of samples
or tissues}
}
\value{
A transformed vector of length \emph{m}
}
\description{
The first step in ROKU algorithm. Normalize the expression data using Tukey's
biweight. Values with \code{NA} are excluded.
}
\references{
Kadota et al (2006). ROKU: a novel method for identification of tissue-specific genes. BMC Genomics, 7:294.
\url{http://www.biomedcentral.com/1471-2105/7/294}
}

