% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{package}
\name{ROKU}
\alias{ROKU}
\alias{ROKU-package}
\title{ROKU}
\usage{
ROKU(m, cores = 1)
}
\arguments{
\item{m}{a \emph{n} by \emph{m} data frame where \emph{n} is the number of genes
and \emph{m} is the number of samples/tissues}

\item{cores}{number of cores to use for parallel processing. Default is 1.}
}
\value{
List:
\itemize{
 \item{entropy}{}
 \item{entropy.pct}{}
 \item{outliers}{}
}
}
\description{
ROKU

A method for identification of tissue-specific genes
}
\details{
See the \href{http://www.biomedcentral.com/1471-2105/7/294#sec4}{Methods} from the manuscript for more details.

General steps:
\enumerate{
 \item Process each vector: x -> x' using Tukey Biweight
 \item Calculate entropy: H(x')
 \item Assign tissues detected as outliers using AIC
 }
}

