% Generated by roxygen2 (4.0.2): do not edit by hand
\name{u_statistic}
\alias{u_statistic}
\title{U statistic}
\usage{
u_statistic(x, outliers, epsilon = 1e-04)
}
\arguments{
\item{x}{}

\item{outliers}{}

\item{epsilon}{}
}
\value{
U statistic (numeric)
}
\description{
U statistic
}
\seealso{
\code{\link{find_tissue_outliers}}
}

